chore(tasks): Migrate util task methods to cli-utils
