+++ repos/cli-utils/src/error/error.js
@@ -55,6 +55,22 @@ const throwNoAction = (task=noOpObj) => {
 }
 
 /**
+ * Prints CLI unknown task when task can not be found
+ * @param {string} task - Invalid passed in task
+ * @param {boolean} unknown - If it's an unknown task
+ *
+ * @returns {void}
+ */
+const throwNoTask = task => {
+  Logger.empty()
+  Logger.error(`Unknown task => ${task}`)
+  Logger.green(`Type "keg help' to see all known tasks.`)
+  Logger.empty()
+
+  throwTaskFailed()
+}
+
+/**
  * Throws task failed error
  * @function
  * @private
@@ -68,6 +84,7 @@ const throwTaskFailed = () => {
 module.exports = {
   throwError,
   throwExitError,
+  throwNoTask,
   throwNoAction,
   throwTaskFailed,
 }
