+++ repos/cli-utils/src/globalConfig/index.js
@@ -1,4 +1,5 @@
 module.exports = {
   ...require('./configHelpers'),
+  ...require('./globalOptions'),
   ...require('./getKegGlobalConfig'),
 }
