+++ repos/cli-utils/src/task/index.js
@@ -1,5 +1,9 @@
 module.exports = {
+  ...require('./executeTask'),
   ...require('./findTask'),
   ...require('./hasHelpArg'),
+  ...require('./parseTaskArgs'),
+  ...require('./runInternalTask'),
   ...require('./sharedOptions'),
+  ...require('./validateTask'),
 }
