+++ src/tasks/base/build.js
@@ -1,5 +1,4 @@
-const { keyMap } = require('@keg-hub/jsutils')
-const { runInternalTask } = require('KegUtils/task/runInternalTask')
+const { runInternalTask } = require('@keg-hub/cli-utils')
 const { mergeTaskOptions } = require('KegUtils/task/options/mergeTaskOptions')
 
 /**
