+++ src/tasks/docker/compose/up.js
@@ -1,9 +1,8 @@
-const { Logger } = require('@keg-hub/cli-utils')
 const { pickKeys } = require('@keg-hub/jsutils')
 const { logVirtualUrl } = require('KegUtils/log')
 const { spawnCmd } = require('@keg-hub/spawn-cmd')
 const { buildContainerContext } = require('KegUtils/builders')
-const { runInternalTask } = require('KegUtils/task/runInternalTask')
+const { Logger, runInternalTask } = require('@keg-hub/cli-utils')
 const { throwComposeFailed } = require('KegUtils/error/throwComposeFailed')
 const { mergeTaskOptions } = require('KegUtils/task/options/mergeTaskOptions')
 const { buildComposeCmd } = require('KegUtils/docker/compose/buildComposeCmd')
