+++ src/tasks/docker/package/package.js
@@ -1,10 +1,9 @@
 const docker = require('@keg-hub/docker-lib')
-const { Logger } = require('@keg-hub/cli-utils')
 const { get, noOpObj } = require('@keg-hub/jsutils')
 const { getImgTags } = require('KegUtils/docker/getImgTags')
 const { CONTEXT_TO_CONTAINER } = require('KegConst/constants')
+const { Logger, runInternalTask} = require('@keg-hub/cli-utils')
 const { throwRequired, generalError } = require('KegUtils/error')
-const { runInternalTask } = require('KegUtils/task/runInternalTask')
 const { runBuildAction } = require('KegUtils/package/runBuildAction')
 const { imageFromContainer } = require('KegUtils/package/imageFromContainer')
 const { mergeTaskOptions } = require('KegUtils/task/options/mergeTaskOptions')
