+++ src/tasks/docker/provider/push.js
@@ -1,10 +1,9 @@
 const docker = require('@keg-hub/docker-lib')
 const { ask } = require('@keg-hub/ask-it')
-const { Logger } = require('@keg-hub/cli-utils')
 const { isObj, noOpObj } = require('@keg-hub/jsutils')
 const { getImageRef } = require('KegUtils/docker/getImageRef')
 const { generalError } = require('KegUtils/error/generalError')
-const { runInternalTask } = require('KegUtils/task/runInternalTask')
+const { Logger, runInternalTask } = require('@keg-hub/cli-utils')
 const { throwNoDockerImg } = require('KegUtils/error/throwNoDockerImg')
 const { getImgNameContext } = require('KegUtils/getters/getImgNameContext')
 const { isValidProviderUrl } = require('KegUtils/helpers/isValidProviderUrl')
