+++ src/tasks/git/branch/branch.js
@@ -1,9 +1,8 @@
-const { isNum, exists } = require('@keg-hub/jsutils')
-const { Logger } = require('@keg-hub/cli-utils')
 const { git } = require('@keg-hub/git-lib')
-const { runInternalTask } = require('KegUtils/task/runInternalTask')
-const { generalError } = require('KegUtils/error')
 const { ask } = require('@keg-hub/ask-it')
+const { generalError } = require('KegUtils/error')
+const { isNum, exists } = require('@keg-hub/jsutils')
+const { Logger, runInternalTask } = require('@keg-hub/cli-utils')
 
 /**
  * Get the branch name based on the branch or the params.remove value
