+++ src/tasks/tap/push.js
@@ -1,5 +1,5 @@
 const { get } = require('@keg-hub/jsutils')
-const { runInternalTask } = require('KegUtils/task/runInternalTask')
+const { runInternalTask } = require('@keg-hub/cli-utils')
 const { mergeTaskOptions } = require('KegUtils/task/options/mergeTaskOptions')
 const { updateLocationContext } = require('KegUtils/helpers/updateLocationContext')
