+++ src/utils/builders/buildGlobalTaskAlias.js
@@ -1,6 +1,6 @@
+const { get, isStr } = require('@keg-hub/jsutils')
 const { throwWrap } = require('../error/throwWrap')
-const { executeTask } = require('../task/executeTask')
-const { get, isFunc, isStr } = require('@keg-hub/jsutils')
+const { executeTask } = require('@keg-hub/cli-utils')
 
 /**
  * Checks all global sub-tasks for sub-sub-tasks
