+++ src/utils/services/__tests__/proxyService.js
@@ -1,3 +1,4 @@
+const cliUtils = require('@keg-hub/cli-utils')
 const { docker } = require('KegMocks/libs/docker')
 const globalConfig = global.getGlobalCliConfig()
 
@@ -6,8 +7,8 @@ jest.setMock('@keg-hub/docker-lib', docker)
 const internalTaskMock = jest.fn(args => {
   return args
 })
+cliUtils.runInternalTask = internalTaskMock
 
-jest.setMock('KegUtils/task/runInternalTask', { runInternalTask: internalTaskMock })
 
 const { proxyService } = require('../proxyService')
