+++ src/utils/services/composeService.js
@@ -1,10 +1,9 @@
-const { Logger } = require('@keg-hub/cli-utils')
 const { mutagenService } = require('./mutagenService')
 const { getServiceArgs } = require('./getServiceArgs')
 const { buildContainerSync } = require('./syncService')
-const { runInternalTask } = require('../task/runInternalTask')
 const { buildExecParams } = require('../docker/buildExecParams')
 const { getContainerCmd } = require('../docker/getContainerCmd')
+const { Logger, runInternalTask } = require('@keg-hub/cli-utils')
 const { KEG_DOCKER_EXEC, KEG_EXEC_OPTS } = require('KegConst/constants')
 const { checkEnvConstantValue } = require('KegUtils/helpers/checkEnvConstantValue')
 const { get, isArr, set, noOpObj, deepMerge, exists } = require('@keg-hub/jsutils')
