+++ src/utils/services/stopService.js
@@ -1,6 +1,5 @@
-const { Logger } = require('@keg-hub/cli-utils')
 const { getServiceArgs } = require('./getServiceArgs')
-const { runInternalTask } = require('KegUtils/task/runInternalTask')
+const { Logger, runInternalTask } = require('@keg-hub/cli-utils')
 const { checkEnvConstantValue } = require('KegUtils/helpers/checkEnvConstantValue')
 
 /**
