+++ src/utils/task/index.js
@@ -1,10 +1,5 @@
 module.exports = {
-  ...require('./convertParamsToEnvs'),
-  ...require('./executeTask'),
   ...require('./findTask'),
   ...require('./getTask'),
-  ...require('./globalOptions'),
   ...require('./options'),
-  ...require('./runInternalTask'),
-  ...require('./validateTask'),
 }
